fix: move TaskEdge routes to tasks.py, correct DELETE path
