+++ src/ainrf/api/routes/tasks.py
@@ -10,6 +10,9 @@ from fastapi.responses import StreamingResponse
 from ainrf.api.schemas import (
     TaskCreateRequest,
     TaskDetailResponse,
+    TaskEdgeCreateRequest,
+    TaskEdgeListResponse,
+    TaskEdgeResponse,
     TaskListResponse,
     TaskOutputEventResponse,
     TaskOutputListResponse,
@@ -26,6 +29,7 @@ from ainrf.task_harness import (
 from ainrf.workspaces import WorkspaceNotFoundError
 
 router = APIRouter(prefix="/tasks", tags=["tasks"])
+task_edges_router = APIRouter(tags=["task-edges"])
 
 
 def _get_task_harness_service(request: Request) -> TaskHarnessService:
@@ -257,3 +261,63 @@ async def cancel_task(task_id: str, request: Request) -> TaskSummaryResponse:
     except Exception as exc:
         raise _translate_task_error(exc) from exc
     return TaskSummaryResponse.model_validate(_serialize_task_summary(task))
+
+
+@task_edges_router.get("/projects/{project_id}/task-edges", response_model=TaskEdgeListResponse)
+async def list_task_edges(project_id: str, request: Request) -> TaskEdgeListResponse:
+    service = _get_task_harness_service(request)
+    try:
+        edges = service.get_task_edges(project_id)
+    except Exception as exc:
+        raise _translate_task_error(exc) from exc
+    return TaskEdgeListResponse.model_validate(
+        {
+            "items": [
+                {
+                    "edge_id": edge.edge_id,
+                    "project_id": edge.project_id,
+                    "source_task_id": edge.source_task_id,
+                    "target_task_id": edge.target_task_id,
+                    "created_at": edge.created_at.isoformat(),
+                }
+                for edge in edges
+            ]
+        }
+    )
+
+
+@task_edges_router.post(
+    "/projects/{project_id}/task-edges",
+    response_model=TaskEdgeResponse,
+    status_code=status.HTTP_201_CREATED,
+)
+async def create_task_edge(
+    project_id: str, payload: TaskEdgeCreateRequest, request: Request
+) -> TaskEdgeResponse:
+    service = _get_task_harness_service(request)
+    try:
+        edge = service.create_task_edge(
+            project_id=project_id,
+            source_task_id=payload.source_task_id,
+            target_task_id=payload.target_task_id,
+        )
+    except Exception as exc:
+        raise _translate_task_error(exc) from exc
+    return TaskEdgeResponse.model_validate(
+        {
+            "edge_id": edge.edge_id,
+            "project_id": edge.project_id,
+            "source_task_id": edge.source_task_id,
+            "target_task_id": edge.target_task_id,
+            "created_at": edge.created_at.isoformat(),
+        }
+    )
+
+
+@task_edges_router.delete("/task-edges/{edge_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_task_edge(edge_id: str, request: Request) -> None:
+    service = _get_task_harness_service(request)
+    try:
+        service.delete_task_edge(edge_id)
+    except Exception as exc:
+        raise _translate_task_error(exc) from exc
